move xarray exceptions into `xarray.errors` module
